Split out code into files and added more checks
